Fix overflow issue of p-ext multiply instructions (#1053)

+++ riscv/insns/smul8.h
@@ -1,3 +1,3 @@
 P_MUL_LOOP(8, {
-  pd = ps1 * ps2;
+  pd = (int16_t)ps1 * (int16_t)ps2;
 })
