+++ riscv/insns/smulx16.h
@@ -1,3 +1,3 @@
 P_MUL_CROSS_LOOP(16, {
-  pd = ps1 * ps2;
+  pd = (int32_t)ps1 * (int32_t)ps2;
 })
